combine sidebar components in a single one
