--- plugins/self-service/src/components/CollectionsSidebarItem/index.ts
@@ -1,1 +1,0 @@
-export { CollectionsSidebarItem } from './CollectionsSidebarItem';
