--- plugins/self-service/src/components/EEBuilderSidebarItem/index.ts
@@ -1,1 +1,0 @@
-export { EEBuilderSidebarItem } from './EEBuilderSidebarItem';
