--- plugins/self-service/src/components/GitRepositoriesSidebarItem/index.ts
@@ -1,1 +1,0 @@
-export { GitRepositoriesSidebarItem } from './GitRepositoriesSidebarItem';
