--- plugins/self-service/src/components/PermissionGatedSidebarItem/index.ts
@@ -1,1 +1,0 @@
-export { PermissionGatedSidebarItem } from './PermissionGatedSidebarItem';
