+++ plugins/self-service/src/components/SidebarItems/index.ts
@@ -1,0 +1,6 @@
+export {
+  PermissionGatedSidebarItem,
+  EEBuilderSidebarItem,
+  CollectionsSidebarItem,
+  GitRepositoriesSidebarItem,
+} from './SidebarItems';
