+++ plugins/self-service/src/index.ts
@@ -3,9 +3,11 @@ export * from './plugin';
 export { SignInPage } from './components/SignInPage';
 export { LandingPage } from './components/LandingPage';
 export { TaskList } from './components/TaskList';
-export { EEBuilderSidebarItem } from './components/EEBuilderSidebarItem';
-export { CollectionsSidebarItem } from './components/CollectionsSidebarItem';
-export { GitRepositoriesSidebarItem } from './components/GitRepositoriesSidebarItem';
+export {
+  EEBuilderSidebarItem,
+  CollectionsSidebarItem,
+  GitRepositoriesSidebarItem,
+} from './components/SidebarItems';
 export { AAPTokenFieldExtension } from './components/Scaffolder/AAPTokenField/extensions';
 export { AAPResourcePickerExtension } from './components/Scaffolder/AAResourcePicker';
 export { BaseImagePickerFieldExtension } from './components/Scaffolder/BaseImagePicker/extensions';
